requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==4.2.1
django-cors-headers==4.0.0
google-auth==2.19.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
Pillow==10.0.0
psycopg==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
zope.interface==6.0
